requirements.txt
supabase==2.15.2
python-dotenv==1.0.1
fastapi==0.115.14
uvicorn==0.35.0
requests
pydantic[email]
